+++ src/core/component/functional/index.ts
@@ -14,4 +14,5 @@
 export * from 'core/component/functional/const';
 export * from 'core/component/functional/fake-ctx';
 export * from 'core/component/functional/vnode';
+export * from 'core/component/functional/helpers';
 export * from 'core/component/functional/interface';