+++ src/core/component/index.ts
@@ -19,6 +19,7 @@ export * from 'core/component/watch';
 
 export * from 'core/component/register';
 export * from 'core/component/reflection';
+export * from 'core/component/method';
 
 export * from 'core/component/event';
 export * from 'core/component/render';