+++ src/core/component/render-function/index.ts
@@ -192,7 +192,7 @@ export function wrapRender(meta: ComponentMeta): RenderFunction {
 										return returnEls([]);
 									}
 
-									Object.set(ctx, 'hook', 'beforeUpdate');
+									ctx.hook = 'beforeUpdate';
 									Object.set(ctx, 'renderGroup', desc?.renderGroup);
 
 									for (let o = forEach(iterable, forEachCb), i = 0; i < o.length; i++) {
@@ -258,7 +258,7 @@ export function wrapRender(meta: ComponentMeta): RenderFunction {
 									runHook('beforeUpdated', ctx, desc).catch(stderr);
 
 									resolveRefs(ctx);
-									Object.set(ctx, 'hook', baseHook);
+									ctx.hook = baseHook;
 
 									return returnEls(els);
 								}