+++ src/core/component/directives/update-on/CHANGELOG.md
@@ -9,6 +9,21 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :boom: Breaking Change
+
+* Deprecate `listener` and `once` parameters
+
+#### :rocket: New Feature
+
+* Added support of watchers
+* Improved API
+
+#### :memo: Documentation
+
+* Improved documentation
+
 ## v3.0.0-rc.112 (2020-12-18)
 
 #### :rocket: New Feature